More code refactoring and adding support for literals.
