+++ src/System.Text.JsonLab/System/Text/Json/Utf8JsonWriter.WriteProperties.Number.cs
@@ -1,0 +1,16 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Buffers;
+using System.Buffers.Text;
+using System.Diagnostics;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+namespace System.Text.JsonLab
+{
+    public ref partial struct Utf8JsonWriter2
+    {
+    }
+}
